Implemented universal library for Jackson and Play JSON libraries
