--- dependencies.sbt
@@ -1,4 +1,0 @@
-libraryDependencies += "org.scala-lang.modules" %% "scala-parser-combinators" % "1.1.2"
-libraryDependencies += "org.scalatest" %% "scalatest" % "3.0.5" % "test"
-libraryDependencies += "org.scalacheck" %% "scalacheck" % "1.14.0" % "test"
-libraryDependencies += "com.typesafe.play" %% "play-json" % "2.7.4"
