+++ jackson/build.sbt
@@ -1,0 +1,1 @@
+name := "jsonpath-jackson"
