+++ jackson/dependencies.sbt
@@ -1,0 +1,3 @@
+libraryDependencies += "com.fasterxml.jackson.core" % "jackson-databind" % "2.9.9"
+libraryDependencies += "org.scalatest" %% "scalatest" % "3.0.5" % "test"
+libraryDependencies += "org.scalacheck" %% "scalacheck" % "1.14.0" % "test"
