+++ jackson/src/test/scala/io/gatling/jsonpath/jackson/JsonPathSpec.scala
@@ -1,0 +1,29 @@
+/*
+ * Copyright 2011-2019 GatlingCorp (https://gatling.io)
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *  http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package io.gatling.jsonpath.jackson
+
+import com.fasterxml.jackson.databind.{ JsonNode, ObjectMapper }
+import io.gatling.jsonpath.{ BaseJsonPath, BaseJsonPathSpec }
+
+class JsonPathSpec extends BaseJsonPathSpec[JsonNode] {
+  override def parseJson(s: String): JsonNode = JsonPathSpec.mapper.readValue(s, classOf[JsonNode])
+  override def JsonPath: BaseJsonPath[JsonNode] = io.gatling.jsonpath.jackson.JsonPath
+}
+
+object JsonPathSpec {
+  val mapper: ObjectMapper = new ObjectMapper()
+}
