+++ play/dependencies.sbt
@@ -1,0 +1,3 @@
+libraryDependencies += "com.typesafe.play" %% "play-json" % "2.7.4"
+libraryDependencies += "org.scalatest" %% "scalatest" % "3.0.5" % "test"
+libraryDependencies += "org.scalacheck" %% "scalacheck" % "1.14.0" % "test"
