+++ play/src/test/scala/io/gatling/jsonpath/play/JsonPathSpec.scala
@@ -1,0 +1,25 @@
+/*
+ * Copyright 2011-2019 GatlingCorp (https://gatling.io)
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *  http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package io.gatling.jsonpath.play
+
+import io.gatling.jsonpath.{ BaseJsonPath, BaseJsonPathSpec }
+import play.api.libs.json.{ JsValue, Json }
+
+class JsonPathSpec extends BaseJsonPathSpec[JsValue] {
+  override def parseJson(s: String): JsValue = Json.parse(s)
+  override def JsonPath: BaseJsonPath[JsValue] = io.gatling.jsonpath.play.JsonPath
+}
